backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
nltk==3.8.1
spacy==3.6.1
reportlab==4.0.4
jinja2==3.1.2
werkzeug==2.3.7
gunicorn==21.2.0
huggingface_hub==0.17.3
transformers==4.35.2